Add lock versions for concurrency control
This adds in the Publishing API optimistic locking approach [1]. This
has been added to ensure that the content in the Publishing API does not
change state during an update to live or draft. Without this we run some
risks if this tool receives concurrent requests. For example, if we were
updating live and draft at the same time, a put-content request of draft
content could be pushed live by mistake due to competing requests.

I made a deliberate decision to look up the lock version id before we
pull data from Contentful. I felt that this would ensure we captured the
exact contentful state after the current version in the Publishing API,
whereas I felt there was some ambiguity if this was pulled before we
looked up the Publishing API version (i.e. could it be now stale?)

Note for content that doesn't already exist we can set a value of "0"
and the Publishing API will reject this if a version already exists (but
will ignore it if there is no edition).

The Publishing API seems to be particular that it needs this
"previous_version" parameter sent as a string - otherwise it rejects
with 422.

[1]: https://github.com/alphagov/publishing-api/blob/main/docs/api.md#optimistic-locking-previous_version