Make content config aware of spaces
Contentful allows an organisation to have multiple spaces, which allow
for distinct content to be managed. It is currently unclear whether
GOV.UK will make use of multiple spaces. This adds space_id in as part
of the config for a piece of content to allow for using them if we want
to.