Combine the updater classes and add retry handling
This combines the LiveUpdater and DraftUpdater classes as they had a
large amount of repetition. It adds retry handling for Publishing API
conflicts as we expect there will be situations when there are
concurrent requests leading to competing attempts to update the
Publishing API (for example if you publish something on Contentful you
get a webhook for save and publish).

Finally this updates the method signatures in ContentState to
differentiate them from the Updater class.