Include space_id with ids to recognise entity
In Contentful an id is made unique by 4 parts: space id, environment id,
entity type, and the entity's id. This updates our handling of this for
greater uniqueness by including space.

I've purposefully chosen not to include environment in the id as we
intend to allow this as something devs can change to explore their own
changes.