Be aware of contentful environment
Contentful allows branching off of the main environment (master) for
development exploration. We don't want to intercept webhooks for those
environments so this adds code to reject them and allow the developers
to configure the environment if they wish.

Worth noting that by default a Contentful webhook will be for the master
environment, so this changes is slightly paranoid.