Include maximum depth of results for entry lookups
This saves having to make additional requests to Contentful by returning
all the nested objects in the response.

10 is the maximum number that Contentful will accept. As we recursively
look up the entire tree we expect to always want the whole depth.

We may also have to add include to the resolve lookups within
ContentPayload should we find that there are deep graphs to traverse.