Use VCR to store Contentful HTTP response
This is a replacement for the rather crude homebrew webmock fixture for
a Contentful API response.

VCR is a gem that allows recording and replaying HTTP responses in
tests. Integrating proved a little awkward as you need credentials to
access Contentful and I don't want to couple the configuration to a
specific content item.

In order to deal with this I've set-up a rake task that can be used to
generate the VCR cassette based on the top item in ContentConfig. A
helper method is then used to match any requests to contentful.com to
return this (we expect two: a draft and a live one).

VCR seems to expect you to use it for all HTTP requests, which is a bit
awkward as GDS API Adapters makes quite substantial use of WebMock. In
order to deal with this I've configured only tests with vcr: true to use
VCR.

I used a Contentful entry that makes use of lots of different types of
data fields. It's unlikely this item will be outside of basic testing so
it will likely be replaced at some point when there is reason to believe
the Contentful API responses have changed.