Add tests for PublishingApi::ContentState
To cover all the code paths these tests are a little contrived. I've
aimed to explore full coverage with them.

I noticed that we didn't handle the scenario of a 404 when trying to
access a published item that doesn't exist, so we now handle that. This
is a very unlikely scenario as the Publishing API will have just given
us data to indicate it does exist.