Add tests for PublishingApi::ContentPayload
Writing the tests revealed a few areas for amends in the code.

These tests were quite tricky to write as it relies on Contentful models
producing expected results. It's hard to test these exactly as you
either need to set-up a ton of webmock for Contentful API responses or
use test doubles for the classes. I went for the latter approach as it
was simpler. The Contentful resource classes rely on dynamic methods
which make them a pain to mock (and it's also a bit annoying for testing
compatibility).

I'm hoping these tests will be sufficient as the real models are used in
the integration test. However if we have problems with API changes or
Contentful gem we may need to refactor them.