Add healthcheck endpoints
This is a step towards integrating this with GOV.UK.

The structure of the contentful check is based on:
https://github.com/alphagov/govuk_app_config/blob/main/docs/healthchecks.md

The Contentful healthcheck involves checking whether it can make the API
call to the space endpoint, which seemed to be the most generic endpoint
I could find.

There are a couple of risks of this: 1) looking up the space seems to be
surprisingly slow - around 250ms, so if we had a bunch of spaces this
could get slow (as an aside I don't why it's slow it should be pretty
simple) and 2) if we do the healthcheck too frequently it could use up
API rate-limiting, hopefully it's infrequent enough for this to be a
problem.