Configure Logstash logging for production
When running in GOV.UK hosting we expect applications to produce
logstash formatted output. Adding this rack-logstasher gem configures
the application to do this.

To set this up, we need to first disable Sinatra logging, which defaults
to using Rack::CommonLogger which produces Apache formatted logs to
STDOUT. Then we need to configure the Racklogger, which adds an entry of
govuk_request_id.

With this configuration running on Unicorn we end up with Unicorn
outputting unformatted log entries to STDERR (unicorn set-up) and just
the JSON formatted request entries to STDOUT. We also have access to a
logger in requests which will also log to STDERR.

Should a request produce an exception, the stacktrace will be written to
STDERR and STDOUT will just have a JSON formatted entry of a 500
response.

I tested this configuration also with Puma, this didn't work as well.
Puma will write it's status information to stdout and there won't be
access to a logger for requests.

To resolve this I'd suggest starting puma in silent mode (puma -s) and
adding a hook to set-up a logger of:

```
before { env["rack.logger"] = Logger.new($stderr) }
```
which will then be available in requests.