Discard drafts when unpublishing
The Publishing API will reject an unpublish if a draft is present if
this is set. As it's quite easy to have an active draft with this
process I've set this to true. It also seems quite unlikely that someone
would run this command with a concern for any drafts.